rename step1.js -> index.js
